requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.0
